+++ src/core/scenario.jl
@@ -107,6 +107,8 @@ function Scenario(d::Dict; flex_hvac_from_json=false)
                 push!(pvs, PV(
                     ; dictkeys_tosymbols(pv)...,
                     off_grid_flag = settings.off_grid_flag, 
+                    sector = site.sector,
+                    federal_procurement_type = site.federal_procurement_type,
                     latitude = site.latitude,
                     electric_load_annual_kwh = electric_load_annual_kwh,
                     site_land_acres = site.land_acres,
@@ -117,6 +119,8 @@ function Scenario(d::Dict; flex_hvac_from_json=false)
             push!(pvs, PV(
                 ; dictkeys_tosymbols(d["PV"])..., 
                 off_grid_flag = settings.off_grid_flag, 
+                sector = site.sector,
+                federal_procurement_type = site.federal_procurement_type,
                 latitude = site.latitude,
                 electric_load_annual_kwh = electric_load_annual_kwh,
                 site_land_acres = site.land_acres,
@@ -131,12 +135,18 @@ function Scenario(d::Dict; flex_hvac_from_json=false)
         financial = Financial(; dictkeys_tosymbols(d["Financial"])...,
                                 latitude=site.latitude, longitude=site.longitude, 
                                 off_grid_flag = settings.off_grid_flag,
-                                include_health_in_objective = settings.include_health_in_objective
+                                include_health_in_objective = settings.include_health_in_objective,
+                                sector = site.sector,
+                                federal_procurement_type = site.federal_procurement_type,
+                                federal_sector_state = site.federal_sector_state
                             )
     else
         financial = Financial(; latitude=site.latitude, longitude=site.longitude,
                                 off_grid_flag = settings.off_grid_flag,
-                                include_health_in_objective = settings.include_health_in_objective
+                                include_health_in_objective = settings.include_health_in_objective,
+                                sector = site.sector,
+                                federal_procurement_type = site.federal_procurement_type,
+                                federal_sector_state = site.federal_sector_state
                             )
     end
 
@@ -185,25 +195,22 @@ function Scenario(d::Dict; flex_hvac_from_json=false)
         
     storage_structs = Dict{String, AbstractStorage}()
     if haskey(d,  "ElectricStorage")
-        storage_dict = dictkeys_tosymbols(d["ElectricStorage"])
-        storage_dict[:off_grid_flag] = settings.off_grid_flag
+        storage_dict = d["ElectricStorage"]
+        storage_dict["off_grid_flag"] = settings.off_grid_flag
     else
-        storage_dict = Dict(:max_kw => 0.0) 
+        storage_dict = Dict("max_kw" => 0.0)
     end
-    storage_structs["ElectricStorage"] = ElectricStorage(storage_dict, financial)
+    storage_structs["ElectricStorage"] = ElectricStorage(storage_dict, financial, site)
     # TODO stop building ElectricStorage when it is not modeled by user 
     #       (requires significant changes to constraints, variables)
     if haskey(d, "HotThermalStorage")
-        params = HotThermalStorageDefaults(; dictkeys_tosymbols(d["HotThermalStorage"])...)
-        storage_structs["HotThermalStorage"] = HotThermalStorage(params, financial, settings.time_steps_per_hour)
+        storage_structs["HotThermalStorage"] = HotThermalStorage(d["HotThermalStorage"], financial, site, settings.time_steps_per_hour)
     end
     if haskey(d, "HighTempThermalStorage")
-        params = HighTempThermalStorageDefaults(; dictkeys_tosymbols(d["HighTempThermalStorage"])...)
-        storage_structs["HighTempThermalStorage"] = HighTempThermalStorage(params, financial, settings.time_steps_per_hour)
+        storage_structs["HighTempThermalStorage"] = HighTempThermalStorage(d["HighTempThermalStorage"], financial, site, settings.time_steps_per_hour)
     end
     if haskey(d, "ColdThermalStorage")
-        params = ColdThermalStorageDefaults(; dictkeys_tosymbols(d["ColdThermalStorage"])...)
-        storage_structs["ColdThermalStorage"] = ColdThermalStorage(params, financial, settings.time_steps_per_hour)
+        storage_structs["ColdThermalStorage"] = ColdThermalStorage(d["ColdThermalStorage"], financial, site, settings.time_steps_per_hour)
     end
     storage = Storage(storage_structs)
 
@@ -226,6 +233,7 @@ function Scenario(d::Dict; flex_hvac_from_json=false)
 
     if haskey(d, "Wind")
         wind = Wind(; dictkeys_tosymbols(d["Wind"])..., off_grid_flag=settings.off_grid_flag,
+                    sector = site.sector, federal_procurement_type = site.federal_procurement_type,
                     average_elec_load=sum(electric_load.loads_kw) / length(electric_load.loads_kw))
     else
         wind = Wind(; max_kw=0)
@@ -411,11 +419,15 @@ function Scenario(d::Dict; flex_hvac_from_json=false)
                     avg_boiler_fuel_load_mmbtu_per_hour = avg_boiler_fuel_load_mmbtu_per_hour,
                     existing_boiler = existing_boiler,
                     electric_load_series_kw = electric_load.loads_kw,
-                    year = electric_load.year)
+                    year = electric_load.year,
+                    sector = site.sector,
+                    federal_procurement_type = site.federal_procurement_type)
         else # Only if modeling CHP without heating_load and existing_boiler (for prime generator, electric-only)
             chp = CHP(d["CHP"],
                     electric_load_series_kw = electric_load.loads_kw,
-                    year = electric_load.year)
+                    year = electric_load.year,
+                    sector = site.sector,
+                    federal_procurement_type = site.federal_procurement_type)
         end
         chp_prime_mover = chp.prime_mover
     end
@@ -831,7 +843,15 @@ function Scenario(d::Dict; flex_hvac_from_json=false)
                     pop!(ghp_inputs_removed_ghpghx_params, param)
                 end
             end                    
-            append!(ghp_option_list, [GHP(ghpghx_response, ghp_inputs_removed_ghpghx_params)])
+            append!(
+                ghp_option_list, 
+                [GHP(
+                    ghpghx_response, 
+                    ghp_inputs_removed_ghpghx_params;
+                    sector = site.sector,
+                    federal_procurement_type = site.federal_procurement_type
+                )]
+            )
             # Print out ghpghx_response for loading into a future run without running GhpGhx.jl again
             # open("scenarios/ghpghx_response.json","w") do f             
         end
@@ -848,7 +868,15 @@ function Scenario(d::Dict; flex_hvac_from_json=false)
                     ghp_inputs_removed_ghpghx_responses["is_ghx_hybrid"] = true
                 end
             end
-            append!(ghp_option_list, [GHP(ghpghx_response, ghp_inputs_removed_ghpghx_responses)])
+            append!(
+                ghp_option_list, 
+                [GHP(
+                    ghpghx_response, 
+                    ghp_inputs_removed_ghpghx_responses;
+                    sector = site.sector,
+                    federal_procurement_type = site.federal_procurement_type
+                )]
+            )
         end
     end
 
@@ -858,9 +886,13 @@ function Scenario(d::Dict; flex_hvac_from_json=false)
             total_fuel_heating_load_mmbtu_per_hour = (space_heating_load.loads_kw + dhw_load.loads_kw + process_heat_load.loads_kw) / existing_boiler.efficiency / KWH_PER_MMBTU
             avg_boiler_fuel_load_mmbtu_per_hour = sum(total_fuel_heating_load_mmbtu_per_hour) / length(total_fuel_heating_load_mmbtu_per_hour)
             steam_turbine = SteamTurbine(d["SteamTurbine"];  
-                                        avg_boiler_fuel_load_mmbtu_per_hour = avg_boiler_fuel_load_mmbtu_per_hour)
+                                        avg_boiler_fuel_load_mmbtu_per_hour = avg_boiler_fuel_load_mmbtu_per_hour,
+                                        sector = site.sector, 
+                                        federal_procurement_type = site.federal_procurement_type)
         else
-            steam_turbine = SteamTurbine(d["SteamTurbine"])
+            steam_turbine = SteamTurbine(d["SteamTurbine"]; 
+                                        sector = site.sector, 
+                                        federal_procurement_type = site.federal_procurement_type)
         end
     end